// src/app/core/models/mapping-row.model.ts
/* tslint:disable */
import { MappingDestination } from "./mapping-destination.model";
import { MappingSource } from "./mapping-source.model";

export type MappingRow = {
  auto_mapped: boolean;
  ccc_value?: string;
  fyle_value: string;
  qbo_value: string;
  source: MappingSource;
  destination: MappingDestination;
};